FEAT : add quiz, organize directory
