+++ sendkite/src/main/java/modern/domain/Color.java
@@ -1,4 +1,4 @@
-package modern;
+package modern.domain;
 
 public enum Color {
     RED,
