+++ sendkite/src/main/java/modern/util/formatter/AppleFormatter.java
@@ -1,0 +1,7 @@
+package modern.util.formatter;
+
+import modern.domain.Apple;
+
+public interface AppleFormatter {
+    String accept(Apple a);
+}
