+++ sendkite/src/main/java/modern/util/formatter/ApplePrintFormatter.java
@@ -1,0 +1,12 @@
+package modern.util.formatter;
+
+import modern.domain.Apple;
+
+public class ApplePrintFormatter implements AppleFormatter{
+
+    @Override
+    public String accept(Apple apple) {
+        String characteristic = apple.getWeight() > 150 ? "heavy" : "light";
+        return "A " + characteristic + " " + apple.getColor() + " apple" ;
+    }
+}
