+++ sendkite/src/main/java/modern/util/predicator/AppleHeavyWeightPredicate.java
@@ -1,4 +1,6 @@
-package modern;
+package modern.util.predicator;
+
+import modern.domain.Apple;
 
 public class AppleHeavyWeightPredicate implements ApplePredicate {
