+++ sendkite/src/main/java/modern/util/predicator/ApplePredicate.java
@@ -1,4 +1,6 @@
-package modern;
+package modern.util.predicator;
+
+import modern.domain.Apple;
 
 public interface ApplePredicate {
     boolean test (Apple apple);
